Add action.abort()
